Add create new topic page; display post list and last commented.
* Change to use IResource for resource interfaces.
* Add column property for retrieving posts count and latest posts.
* Add basic topic form rendering.
* Some slight CSS reorganization.

This commit should be a bare minimum to make posting new topic possible.
Posts count and latest post timestamp are retrieved using subquery
instead of storing directly in a database column and are undeferred by
default (i.e. not query until it is accessed). It is only deferred for
listing operation to reduce query count.