Restrict posting if post limit is exceeded or topic is locked.
When topic is locked or archived, it is now no longer possible to create
any more post to that topic. User will be presented with an error page
if topic status changed while he's posting.

Checking is inside retry block in a view so post will properly fail when
topic status changed while retry is still running.