Remove Bootstrap components.
Switching over to a home-made CSS to keep the interface lightweight.
There are many parts of Bootstrap that we don't use and probably
won't in a foreseeable future.

This switch also give us a little more flexible in choosing which
JavaScript framework to use; the first choice being standard-stack
enderjs.