Add timestamp to all assets path for immediate cache expiration.
When HTTP Expires header is used for optimizing page loading, the file
will be cached by the browser or proxy server and may cause assets to
become out of sync when the app is updated. A query string containing
timestamp of asset file is used here to force expiration of cache when
the asset file is updated.

This is, however, is non-optimal since Squid and older browsers refuse
caching anything with ? in its resource URL regardless of HTTP Expires
cookie set.