Add scroll to bottom and scroll to top to top-right and bottom-right
This is mainly for mobile. In a long topic, there is no easy way to
scroll to bottom without tiresome manual scrolling. The new button
on top right of all pages now acts as a scroller, when user tap
on it he will be sent directly to bottom of page. The reverse
equivalent is added in bottom right to page header.

Per request: https://fanboi.ch/lounge/1/42-71