Make popover no longer dismiss if user try to mouse over it.
Using delay trick so when user could mouse over the popover. Once user
move mouse off the link target, the popup won't be dismissed until 100ms
later. If user mouse happens to enter the popover area during the delay
then the dismiss is cancelled.

Once user mouse enter the popover area, an event is added to the popover
to detect that user mouse has leave the popover area and dismiss the
popover accordingly.