Slightly more reliable way to remove popover.
The previous implementation require target.parentNode to be a direct
parent of popover in order to remove the popover. This cause popover
to be unremovable sometimes in situation where two popovers are
nested and user move his mouse out of *both* popovers. This change
fixes that by using parentNode of the popover node instead of
target.parentNode.

Additionally, this commit also allows recursive quote: previously
recursion wasn't working since we rely on cloneNode but that node
may already contains popover. In order to make it work, instead of
inserting cloned node into nodeList right away, we clean it up first
then push it to nodeList.

Some styling fix is also done, namely in "small desktop" layout
padding is no longer added to the side (due to container).