Convert remaining POST actions from traversal to routes.
This completes traversal to routes conversion on view side of things
and re-enabled ability to post new topic and reply to a topic.

Implementation is roughtly the same, except we now use "l5" as
redirect path after reply to a topic instead of calculating range
from post number returned after inserting row.