Calculate total post count from post number instead of actual count.
If non-latest posts are deleted, post count and post number won't be in
sync if we use actual count. This cause post number to exceed number of
max posts before the thread is locked (since count is smaller than post
number).

This commit also fixes a problem where format_markdown can't accept
None and cause the app to fails.