Move topic posting and topic replying stuff into task queue.
* Setup Akismet key on startup instead of via request.
* Move __VERSION__ to fanboi2.version so task queue could access it.
* No longer requiring a real pyramid.util.Request for fanboi2.utils.

This commit moves all the post and topic creation code to task queue
in order to prevent blocking of worker when a post is made. Worker
need to connect to Akismet, which may cause worker to be blocked up
to few seconds depending on network condition.

The implement is still incomplete, as there is no feedback after the
posting is done. Only a dumb redirect.