Add fanboi2.utils.Dnsbl class for checking against DNSBL.
Just a simple wrapper class that make DNS requests with the reversed
IP address to the given DNSBL server. DNSBL servers are DNS servers
that returns a fixed format response, usually in 127.0.0.0/8 class
for the given DNS request to determine whether the IP address is
known spammer or open proxy.