Rewrite all of the frontend-related components.
For extensibility, consistency and maintainability.

Extensibility: after this changes, the layout is separated from the
look of the site. It is now possible to create custom themes for
the site with minimum overrides.

Consistency: now the frontend styles uses shared variables for its
color, metric, font-size and line-height values. The overall look
of the site should be more consistent than before.

Maintainbility: SCSS is now being used instead of Stylus as the
former has much bigger community. Few PostCSS modules is chosen to
automatically transform and optimize CSS (for prefixes, cache-busting
hash, and such.)