Make serializer always return date in configured timezones.
In order to make JS components consistent with the static rendered page.

JS component has no way to be aware of server's timezone, but it need
to render the view in consistent way with the rendered page. If the
API return UTC timestamp, we would need to either hard-code the timezone
or put a timezone hint somewhere in the webpage.

Embedding timezone in timestamp returned in API seems cleanest and require
least changes to both server and JS.