Switch to Python 3.5.
As libraries are dropping Python 3.2 support. This change also switched
from PyPy back to CPython since PyPy3.3 is still alpha, and Python 3.5
support is still forthcoming.

Following to this change, libraries are either deprecated, upgraded or
replaced as appropriate. The list includes:

* Replaced IPy with built-in ipaddress module.
* Replaced mock with built-in unittest.mock module.
* Replaced pg8000 with psycopg2.
* Updated Pyramid to 1.7.
* Updated Alembic to 0.8.
* Updated MarkupSafe to latest.
* Updated wtforms to latest.
* Updated coverage to latest.
* Removed Pygments version lock.

Any code that are incompatible with Python 3.5 are also updated.