chore: switch to pip-tools
Going back to the good 'ol Makefile-based workflow as poetry is getting a bit slow
especially when invoking `poetry run` (though this seems to have been improved lately).

Another reason of dropping Poetry is due to Fanboi2 is being deployed using sdist on a
system without Poetry installed, and with setuptools 63 support of pyproject.toml and
`dynamic` configuration block in setuptools 61, it's not possible to use
requirements.txt as a source of truth in `pip install -e .` scenario as well.

Pip-tools is chosen because of the ease of use.

https://todo.sr.ht/~sirn/fanboi2/11